#16: rename photopage.dart to postmodal.dart
